+++ PowerSync/PowerSync.Common/Client/Sync/Bucket/SyncDataBucket.cs
@@ -2,6 +2,7 @@ namespace PowerSync.Common.Client.Sync.Bucket;
 
 using System.Collections.Generic;
 using System.Linq;
+
 using Newtonsoft.Json;
 
 public class SyncDataBucketJSON