+++ PowerSync/PowerSync.Common/Client/Sync/Stream/StreamingSyncTypes.cs
@@ -1,8 +1,9 @@
 namespace PowerSync.Common.Client.Sync.Stream;
 
+using Newtonsoft.Json;
+
 using PowerSync.Common.Client.Sync.Bucket;
 using PowerSync.Common.DB.Crud;
-using Newtonsoft.Json;
 
 public class ContinueCheckpointRequest
 {
@@ -65,6 +66,9 @@ public class StreamingSyncRequest
 
     [JsonProperty("raw_data")]
     public bool RawData { get; set; }
+
+    [JsonProperty("app_metadata")]
+    public Dictionary<string, string>? AppMetadata { get; set; }
 
     [JsonProperty("parameters")]
     public Dictionary<string, object>? Parameters { get; set; }