+++ PowerSync/PowerSync.Maui/SQLite/MAUISQLiteAdapter.cs
@@ -1,6 +1,7 @@
 namespace PowerSync.Maui.SQLite;
 
 using Microsoft.Data.Sqlite;
+
 using PowerSync.Common.MDSQLite;
 
 // iOS specific imports