+++ Tests/PowerSync/PowerSync.Common.IntegrationTests/NodeConnector.cs
@@ -3,11 +3,12 @@
 
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Net.Http;
 using System.Text;
 using System.Text.Json;
 using System.Threading.Tasks;
-using System.IO;
+
 using PowerSync.Common.Client;
 using PowerSync.Common.Client.Connection;
 using PowerSync.Common.DB.Crud;