+++ demos/CommandLine/Models/Supabase/Todos.cs
@@ -1,6 +1,8 @@
 
 using Microsoft.VisualBasic;
+
 using Newtonsoft.Json;
+
 using Supabase.Postgrest.Attributes;
 using Supabase.Postgrest.Models;