+++ demos/CommandLine/NodeConnector.cs
@@ -2,15 +2,17 @@
 
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Net.Http;
 using System.Text;
 using System.Text.Json;
 using System.Threading.Tasks;
-using System.IO;
+
+using CommandLine.Utils;
+
 using PowerSync.Common.Client;
 using PowerSync.Common.Client.Connection;
 using PowerSync.Common.DB.Crud;
-using CommandLine.Utils;
 
 public class NodeConnector : IPowerSyncBackendConnector
 {