+++ demos/MAUITodo/Platforms/iOS/AppDelegate.cs
@@ -5,5 +5,5 @@ namespace MAUITodo;
 [Register("AppDelegate")]
 public class AppDelegate : MauiUIApplicationDelegate
 {
-	protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
+    protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
 }