+++ demos/WPF/MainWindow.xaml.cs
@@ -1,4 +1,5 @@
 ﻿using System.Windows;
+
 using PowersyncDotnetTodoList.ViewModels;
 
 namespace PowersyncDotnetTodoList;